Player Prefab, Controller and Ability Interface
This commit is the fixed and improved version of Hugos' previous code
for the player as well as an imporved prefab object.
It also implements the Ability system, already incorporated into the
player with the first ability, jump!
Commited is also some tweakings to Unity's input system to allow a more
smooth usage from the player.
This also incorporates the BipolarException class to be the Super class
for all of our custom defined exceptions as well as the first exception,
ComponentNotFound