Snap to magnet
Snap implemented,
To tweak it's behavior, change the following variables on the player
prefab:
SnapAngleThereshold - Changes the angle which you get released from the
magnet.
SnapInstanceForce - Force applied when the mouse is iddle
SnapContinuousForce - Force applied when the player is trying to move
the mouse from the initial rotation. Valid values between 0 and 1.